Reuse the dynamic api prefix resolution for loading public/authorized vocabulary detail.
