+++ src/action/AsyncActions.ts
@@ -5,77 +5,51 @@ import {
   asyncActionSuccess,
   asyncActionSuccessWithPayload,
   publishMessage,
-  publishNotification,
+  publishNotification
 } from "./SyncActions";
-import Ajax, {
-  accept,
-  content,
-  contentType,
-  param,
-  params,
-} from "../util/Ajax";
-import { GetStoreState, TermFetchParams, ThunkDispatch } from "../util/Types";
+import Ajax, {accept, content, contentType, param, params,} from "../util/Ajax";
+import {GetStoreState, TermFetchParams, ThunkDispatch} from "../util/Types";
 import Routing from "../util/Routing";
 import Constants from "../util/Constants";
-import Vocabulary, {
-  CONTEXT as VOCABULARY_CONTEXT,
-  VocabularyData,
-} from "../model/Vocabulary";
-import Routes, { Route } from "../util/Routes";
-import { ErrorData } from "../model/ErrorInfo";
-import { AxiosResponse } from "axios";
+import Vocabulary, {CONTEXT as VOCABULARY_CONTEXT, VocabularyData,} from "../model/Vocabulary";
+import Routes, {Route} from "../util/Routes";
+import {ErrorData} from "../model/ErrorInfo";
+import {AxiosResponse} from "axios";
 import * as jsonld from "jsonld";
 import Message from "../model/Message";
 import MessageType from "../model/MessageType";
-import Term, { CONTEXT as TERM_CONTEXT, TermData } from "../model/Term";
-import VocabularyUtils, { IRI, IRIImpl } from "../util/VocabularyUtils";
+import Term, {CONTEXT as TERM_CONTEXT, TermData} from "../model/Term";
+import VocabularyUtils, {IRI, IRIImpl} from "../util/VocabularyUtils";
 import ActionType from "./ActionType";
-import Resource, { ResourceData } from "../model/Resource";
-import RdfsResource, {
-  CONTEXT as RDFS_RESOURCE_CONTEXT,
-  RdfsResourceData,
-} from "../model/RdfsResource";
-import {
-  CONTEXT as TERM_ASSIGNMENTS_CONTEXT,
-  TermAssignments,
-} from "../model/TermAssignments";
+import Resource, {ResourceData} from "../model/Resource";
+import RdfsResource, {CONTEXT as RDFS_RESOURCE_CONTEXT, RdfsResourceData,} from "../model/RdfsResource";
+import {CONTEXT as TERM_ASSIGNMENTS_CONTEXT, TermAssignments,} from "../model/TermAssignments";
 import TermItState from "../model/TermItState";
 import Utils from "../util/Utils";
-import { CONTEXT as DOCUMENT_CONTEXT } from "../model/Document";
-import {
-  Configuration,
-  CONTEXT as CONFIGURATION_CONTEXT,
-} from "../model/Configuration";
+import {CONTEXT as DOCUMENT_CONTEXT} from "../model/Document";
+import {Configuration, CONTEXT as CONFIGURATION_CONTEXT,} from "../model/Configuration";
 import TermitFile from "../model/File";
 import Asset from "../model/Asset";
 import AssetFactory from "../util/AssetFactory";
 import JsonLdUtils from "../util/JsonLdUtils";
-import { Action } from "redux";
+import {Action} from "redux";
 import {
   CONTEXT as TEXT_ANALYSIS_RECORD_CONTEXT,
   TextAnalysisRecord,
   TextAnalysisRecordData,
 } from "../model/TextAnalysisRecord";
-import {
-  CONTEXT as RESOURCE_TERM_ASSIGNMENTS_CONTEXT,
-  ResourceTermAssignments,
-} from "../model/ResourceTermAssignments";
-import {
-  ChangeRecordData,
-  CONTEXT as CHANGE_RECORD_CONTEXT,
-} from "../model/changetracking/ChangeRecord";
+import {CONTEXT as RESOURCE_TERM_ASSIGNMENTS_CONTEXT, ResourceTermAssignments,} from "../model/ResourceTermAssignments";
+import {ChangeRecordData, CONTEXT as CHANGE_RECORD_CONTEXT,} from "../model/changetracking/ChangeRecord";
 import RecentlyModifiedAsset, {
   CONTEXT as RECENTLY_MODIFIED_ASSET_CONTEXT,
   RecentlyModifiedAssetData,
 } from "../model/RecentlyModifiedAsset";
 import NotificationType from "../model/NotificationType";
-import { langString } from "../model/MultilingualString";
-import ValidationResult, {
-  CONTEXT as VALIDATION_RESULT_CONTEXT,
-} from "../model/ValidationResult";
-import { ConsolidatedResults } from "../model/ConsolidatedResults";
-import UserRole, { UserRoleData } from "../model/UserRole";
-import { loadTermCount } from "./AsyncVocabularyActions";
+import {langString} from "../model/MultilingualString";
+import ValidationResult, {CONTEXT as VALIDATION_RESULT_CONTEXT,} from "../model/ValidationResult";
+import {ConsolidatedResults} from "../model/ConsolidatedResults";
+import UserRole, {UserRoleData} from "../model/UserRole";
+import {loadTermCount} from "./AsyncVocabularyActions";
 import {getApiPrefix} from "./ActionUtils";
 
 /*
@@ -140,7 +114,6 @@ export function createVocabulary(vocabulary: Vocabulary) {
 export function loadVocabulary(
   iri: IRI,
   ignoreLoading: boolean = false,
-  apiPrefix: string = Constants.API_PREFIX,
   withValidation = true
 ) {
   const action = {
@@ -152,7 +125,7 @@ export function loadVocabulary(
     }
     dispatch(asyncActionRequest(action, ignoreLoading));
     return Ajax.get(
-      `${apiPrefix}/vocabularies/${iri.fragment}`,
+      `${getApiPrefix(getState())}/vocabularies/${iri.fragment}`,
       param("namespace", iri.namespace)
     )
       .then((data: object) =>
@@ -162,7 +135,7 @@ export function loadVocabulary(
         )
       )
       .then((data: VocabularyData) => {
-        dispatch(loadImportedVocabulariesIntoState(iri, apiPrefix));
+        dispatch(loadImportedVocabulariesIntoState(iri));
         if (withValidation) {
           dispatch(validateVocabulary(iri));
         }
@@ -181,16 +154,15 @@ export function loadVocabulary(
 }
 
 function loadImportedVocabulariesIntoState(
-  vocabularyIri: IRI,
-  apiPrefix: string
+  vocabularyIri: IRI
 ) {
   const action = {
     type: ActionType.LOAD_VOCABULARY_IMPORTS,
   };
-  return (dispatch: ThunkDispatch) => {
+  return (dispatch: ThunkDispatch, getState: GetStoreState) => {
     dispatch(asyncActionRequest(action, true));
     return Ajax.get(
-      `${apiPrefix}/vocabularies/${vocabularyIri.fragment}/imports`,
+      `${getApiPrefix(getState())}/vocabularies/${vocabularyIri.fragment}/imports`,
       param("namespace", vocabularyIri.namespace)
     )
       .then((data) => dispatch(asyncActionSuccessWithPayload(action, data)))
@@ -240,7 +212,7 @@ export function loadResource(iri: IRI) {
   const action = {
     type: ActionType.LOAD_RESOURCE,
   };
-  return (dispatch: ThunkDispatch, getState: GetStoreState) => {
+  return (dispatch: ThunkDispatch) => {
     dispatch(asyncActionRequest(action));
     return Ajax.get(
       Constants.API_PREFIX + "/resources/" + iri.fragment,
