+++ src/action/__tests__/AsyncActions.test.ts
@@ -285,6 +285,56 @@ describe("Async actions", () => {
         expect(loadImportsSuccessAction.payload).toEqual(imports);
       });
     });
+
+    it("uses public API endpoint to load a single vocabulary when user is not logged in", () => {
+      Ajax.get = jest
+          .fn()
+          .mockImplementation(() =>
+              Promise.resolve(require("../../rest-mock/vocabulary"))
+          );
+      Ajax.head = jest.fn().mockResolvedValue({ headers: {} });
+      return Promise.resolve(
+          (store.dispatch as ThunkDispatch)(
+              loadVocabulary({ fragment: "metropolitan-plan" })
+          )
+      ).then(() => {
+        const url = (Ajax.get as jest.Mock).mock.calls[0][0];
+        expect(url).toContain(Constants.PUBLIC_API_PREFIX);
+        const loadSuccessAction: AsyncActionSuccess<Vocabulary> = store
+            .getActions()
+            .find(
+                (a) =>
+                    a.type === ActionType.LOAD_VOCABULARY &&
+                    a.status === AsyncActionStatus.SUCCESS
+            );
+        expect(loadSuccessAction).toBeDefined();
+        expect(
+            VocabularyUtils.create(loadSuccessAction.payload.iri).fragment ===
+            "metropolitan-plan"
+        ).toBeTruthy();
+      });
+    });
+
+    it("uses public API endpoint to load vocabulary's imports as well when user is not logged in", () => {
+      Ajax.get = jest
+          .fn()
+          .mockImplementation(() =>
+              Promise.resolve(require("../../rest-mock/vocabulary"))
+          );
+      return Promise.resolve(
+          (store.dispatch as ThunkDispatch)(
+              loadVocabulary({ fragment: "metropolitan-plan" })
+          )
+      ).then(() => {
+        const loadImportsAction = store
+            .getActions()
+            .find((a) => a.type === ActionType.LOAD_VOCABULARY_IMPORTS);
+        expect(loadImportsAction).toBeDefined();
+        expect((Ajax.get as jest.Mock).mock.calls.length).toBeGreaterThan(2);
+        const url = (Ajax.get as jest.Mock).mock.calls[1][0];
+        expect(url).toContain(Constants.PUBLIC_API_PREFIX);
+      });
+    });
   });
 
   describe("removeVocabulary", () => {
