+++ src/component/file/FileContentDetail.tsx
@@ -20,7 +20,6 @@ import Mask from "../misc/Mask";
 import AppNotification from "../../model/AppNotification";
 import NotificationType from "../../model/NotificationType";
 import { loadAllTerms } from "../../action/AsyncAnnotatorActions";
-import Constants from "../../util/Constants";
 
 interface FileDetailProvidedProps {
   iri: IRI;
@@ -156,7 +155,7 @@ export default connect(
       clearFileContent: () => dispatch(clearFileContent()),
       loadVocabulary: (vocabularyIri: IRI) =>
         dispatch(
-          loadVocabulary(vocabularyIri, true, Constants.API_PREFIX, false)
+          loadVocabulary(vocabularyIri, true, false)
         ),
       fetchTerms: (vocabularyIri: IRI) =>
         dispatch(loadAllTerms(vocabularyIri, true)),
