+++ src/component/public/term/TermDetail.tsx
@@ -10,10 +10,7 @@ import TermItState from "../../../model/TermItState";
 import { ThunkDispatch } from "../../../util/Types";
 import { injectIntl } from "react-intl";
 import TermMetadata from "./TermMetadata";
-import {
-  loadPublicTerm,
-  loadPublicVocabulary,
-} from "../../../action/AsyncPublicViewActions";
+import { loadPublicTerm } from "../../../action/AsyncPublicViewActions";
 import {
   getLocalized,
   getLocalizedPlural,
@@ -23,6 +20,7 @@ import {
   resolveInitialLanguage,
 } from "../../term/TermDetail";
 import WindowTitle from "../../misc/WindowTitle";
+import { loadVocabulary } from "../../../action/AsyncActions";
 
 interface TermDetailProps
   extends CommonTermDetailProps,
@@ -96,7 +94,7 @@ export default connect(
   },
   (dispatch: ThunkDispatch) => {
     return {
-      loadVocabulary: (iri: IRI) => dispatch(loadPublicVocabulary(iri)),
+      loadVocabulary: (iri: IRI) => dispatch(loadVocabulary(iri)),
       loadTerm: (termName: string, vocabularyIri: IRI) =>
         dispatch(loadPublicTerm(termName, vocabularyIri)),
     };
